requirements.txt
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==2.2.16
djangorestframework==3.11.1
Faker==4.1.3
importlib-metadata==1.7.0
ipdb==0.13.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.2.5
win-unicode-console==0.5
zipp==1.2.0
